+++ pocs/utils/rs232.py
@@ -20,7 +20,10 @@ class SerialData(PanBase):
                  port=None,
                  baudrate=115200,
                  threaded=None,
-                 name="serial_data"):
+                 name="serial_data",
+                 open_delay=2.0,
+                 retry_limit=5,
+                 retry_delay=0.5):
         """Init a SerialData instance.
 
         Args:
@@ -30,41 +33,45 @@ class SerialData(PanBase):
                 the device.
             threaded: Obsolete, ignored.
             name: Name of this object.
+            open_delay: Seconds to wait after opening the port.
+            retry_limit: Number of times to try readline() calls in read().
+            retry_delay: Delay between readline() calls in read().
         """
         PanBase.__init__(self)
 
-        try:
-            self.ser = serial.Serial()
-            self.ser.port = port
-            self.ser.baudrate = baudrate
+        if not port:
+            raise ValueError('Must specify port for SerialData')
 
-            self.ser.bytesize = serial.EIGHTBITS
-            self.ser.parity = serial.PARITY_NONE
-            self.ser.stopbits = serial.STOPBITS_ONE
-            self.ser.timeout = 1.0
-            self.ser.xonxoff = False
-            self.ser.rtscts = False
-            self.ser.dsrdtr = False
-            self.ser.write_timeout = False
-            self.ser.open()
+        self.name = name
+        self.retry_limit = retry_limit
+        self.retry_delay = retry_delay
 
-            self.name = name
+        self.ser = serial.serial_for_url(port, do_not_open=True)
 
-            self.logger.debug(
-                'Serial connection set up to {}, sleeping for two seconds'.format(
-                    self.name))
-            time.sleep(2)
-            self.logger.debug('SerialData created')
-        except Exception as err:
-            self.ser = None
-            self.logger.critical('Could not set up serial port {} {}'.format(
-                port, err))
+        # Configure the PySerial class.
+        self.ser.baudrate = baudrate
+        self.ser.bytesize = serial.EIGHTBITS
+        self.ser.parity = serial.PARITY_NONE
+        self.ser.stopbits = serial.STOPBITS_ONE
+        self.ser.timeout = 1.0
+        self.ser.xonxoff = False
+        self.ser.rtscts = False
+        self.ser.dsrdtr = False
+        self.ser.write_timeout = False
+
+        # Properties have been set to reasonable values, ready to open the port.
+        self.ser.open()
+
+        open_delay = max(0.0, float(open_delay))
+        self.logger.debug(
+            'Serial connection set up to %r, sleeping for %r seconds', self.name, open_delay)
+        if open_delay > 0.0:
+            time.sleep(open_delay)
+        self.logger.debug('SerialData created')
 
     @property
     def is_connected(self):
-        """
-        Checks the serial connection on the mount to determine if connection is open
-        """
+        """True if serial port is open, False otherwise."""
         connected = False
         if self.ser:
             connected = self.ser.isOpen()
@@ -72,8 +79,9 @@ class SerialData(PanBase):
         return connected
 
     def connect(self):
-        """ Actually set up the Thread and connect to serial """
-
+        """Actually set up the Thread and connect to serial."""
+        # TODO(jamessynge): Determine if we need this since the serial port is opened
+        # when the instance is created.
         self.logger.debug('Serial connect called')
         if not self.ser.isOpen():
             try:
@@ -89,7 +97,7 @@ class SerialData(PanBase):
         return self.ser.isOpen()
 
     def disconnect(self):
-        """Closes the serial connection
+        """Closes the serial connection.
 
         Returns:
             bool: Indicates if closed or not
@@ -98,9 +106,7 @@ class SerialData(PanBase):
         return not self.is_connected
 
     def write(self, value):
-        """
-            For now just pass the value along to serial object
-        """
+        """Write value (a string) after encoding as bytes."""
         assert self.ser
         assert self.ser.isOpen()
 
@@ -109,35 +115,39 @@ class SerialData(PanBase):
 
         return response
 
-    def read(self):
-        """
-        Reads value using readline
-        If no response is given, delay and then try to read again. Fail after 10 attempts
+    def read(self, retry_limit=None, retry_delay=None):
+        """Reads next line of input using readline.
+
+        If no response is given, delay for retry_delay and then try to read
+        again. Fail after retry_limit attempts.
         """
         assert self.ser
         assert self.ser.isOpen()
 
-        retry_limit = 5
-        delay = 0.5
+        if retry_limit is None:
+            retry_limit = self.retry_limit
+        if retry_delay is None:
+            retry_delay = self.retry_delay
 
         while True and retry_limit:
             response_string = self.ser.readline(self.ser.inWaiting()).decode()
             if response_string > '':
                 break
-            time.sleep(delay)
+            time.sleep(retry_delay)
             retry_limit -= 1
 
         # self.logger.debug('Serial read: {}'.format(response_string))
-
         return response_string
 
     def get_reading(self):
-        """ Get reading from the queue
+        """Read a line and return the timestamp of the read.
 
         Returns:
             str: Item in queue
         """
 
+        # TODO(jamessynge): Consider collecting the time (utc?) after the read completes,
+        # so that long delays reading don't appear to have happened much earlier.
         try:
             ts = time.strftime('%Y-%m-%dT%H:%M:%S %Z', time.gmtime())
             info = (ts, self.read())
@@ -147,14 +157,26 @@ class SerialData(PanBase):
             return info
 
     def clear_buffer(self):
-        """ Clear Response Buffer """
-        count = 0
-        while self.ser.inWaiting() > 0:
-            count += 1
-            self.ser.read(1)
+        """Clear buffered data from connected port/device.
 
-        # self.logger.debug('Cleared {} bytes from buffer'.format(count))
+        Note that Wilfred reports that the input from an Arduino can seriously lag behind
+        realtime (e.g. 10 seconds), and that clear_buffer may exist for that reason (i.e. toss
+        out any buffered input from a device, and then read the next full line, which likely
+        requires tossing out a fragment of a line).
+        """
+        self.ser.reset_input_buffer()
 
     def __del__(self):
-        if self.ser:
+        """Close the serial device on delete.
+
+        This is to avoid leaving a file or device open if there are multiple references
+        to the serial.Serial object.
+        """
+        try:
+            # If an exception is thrown when running __init__, then self.ser may not have
+            # been set, in which case reading that field will generate a NameError.
+            ser = self.ser
+        except NameError:
+            return
+        if ser:
             self.ser.close()